backend/requirements.txt
# Backend Framework
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.9
psycopg2-binary>=2.9.6
alembic>=1.10.3

# Background Processing
celery>=5.2.7
redis>=4.5.4
flower>=1.2.0

# AI/ML
anthropic>=0.5.0
openai>=1.12.0
aiohttp>=3.9.0
boto3>=1.26.115
numpy>=1.24.2
scikit-learn>=1.2.2
pillow>=9.5.0

# Web Crawling
requests>=2.28.2
beautifulsoup4>=4.12.2
selenium>=4.9.0
webdriver-manager>=3.8.6

# Vector Database
qdrant-client>=1.1.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0
